requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
packaging==25.0
pycryptodome==3.23.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
